+++ src/hooks/useConnectors.ts
@@ -1,18 +1,12 @@
 import { Web3ReactHooks } from '@web3-react/core'
 import { Connector } from '@web3-react/types'
-import { ConnectionType, gnosisSafe, injected, network } from 'connectors'
+import { ConnectionType } from 'connectors'
 import { getConnectorForConnectionType, getHooksForConnectionType } from 'connectors/utils'
-import { useEffect, useMemo } from 'react'
+import { useMemo } from 'react'
+import { BACKFILLABLE_WALLETS } from 'state/connection/constants'
 import { useAppSelector } from 'state/hooks'
 
-import { isMobile } from '../../utils/userAgent'
-
-export const BACKFILLABLE_WALLETS = [
-  ConnectionType.COINBASE_WALLET,
-  ConnectionType.WALLET_CONNECT,
-  ConnectionType.INJECTED,
-]
-export const SELECTABLE_WALLETS = [...BACKFILLABLE_WALLETS, ConnectionType.FORTMATIC]
+const SELECTABLE_WALLETS = [...BACKFILLABLE_WALLETS, ConnectionType.FORTMATIC]
 
 interface Connection {
   connector: Connector
@@ -26,7 +20,7 @@ function getConnectionForConnectionType(connectionType: ConnectionType): Connect
   }
 }
 
-export function useConnectors() {
+export default function useConnectors() {
   const selectedWallet = useAppSelector((state) => state.user.selectedWallet)
   return useMemo(() => {
     const connections: Connection[] = []
@@ -49,37 +43,4 @@ export function useConnectors() {
     const web3Connectors: [Connector, Web3ReactHooks][] = connections.map(({ connector, hooks }) => [connector, hooks])
     return web3Connectors
   }, [selectedWallet])
-}
-
-async function connect(connector: Connector) {
-  try {
-    if (connector.connectEagerly) {
-      await connector.connectEagerly()
-    } else {
-      await connector.activate()
-    }
-  } catch (error) {
-    console.debug(`web3-react eager connection error: ${error}`)
-  }
-}
-
-export function useEagerlyConnect() {
-  const selectedWalletBackfilled = useAppSelector((state) => state.user.selectedWalletBackfilled)
-  const selectedWallet = useAppSelector((state) => state.user.selectedWallet)
-
-  const isMetaMask = !!window.ethereum?.isMetaMask
-
-  useEffect(() => {
-    connect(gnosisSafe)
-    connect(network)
-
-    if (isMobile && isMetaMask) {
-      injected.activate()
-    } else if (selectedWallet) {
-      connect(getConnectorForConnectionType(selectedWallet))
-    } else if (!selectedWalletBackfilled) {
-      BACKFILLABLE_WALLETS.map(getConnectorForConnectionType).forEach(connect)
-    }
-    // The dependency list is empty so this is only run once on mount
-  }, []) // eslint-disable-line react-hooks/exhaustive-deps
 }
