+++ src/hooks/useEagerlyConnect.ts
@@ -1,0 +1,40 @@
+import { Connector } from '@web3-react/types'
+import { gnosisSafe, injected, network } from 'connectors'
+import { getConnectorForConnectionType } from 'connectors/utils'
+import { useEffect } from 'react'
+import { BACKFILLABLE_WALLETS } from 'state/connection/constants'
+import { useAppSelector } from 'state/hooks'
+import { isMobile } from 'utils/userAgent'
+
+async function connect(connector: Connector) {
+  try {
+    if (connector.connectEagerly) {
+      await connector.connectEagerly()
+    } else {
+      await connector.activate()
+    }
+  } catch (error) {
+    console.debug(`web3-react eager connection error: ${error}`)
+  }
+}
+
+export default function useEagerlyConnect() {
+  const selectedWalletBackfilled = useAppSelector((state) => state.user.selectedWalletBackfilled)
+  const selectedWallet = useAppSelector((state) => state.user.selectedWallet)
+
+  const isMetaMask = !!window.ethereum?.isMetaMask
+
+  useEffect(() => {
+    connect(gnosisSafe)
+    connect(network)
+
+    if (isMobile && isMetaMask) {
+      injected.activate()
+    } else if (selectedWallet) {
+      connect(getConnectorForConnectionType(selectedWallet))
+    } else if (!selectedWalletBackfilled) {
+      BACKFILLABLE_WALLETS.map(getConnectorForConnectionType).forEach(connect)
+    }
+    // The dependency list is empty so this is only run once on mount
+  }, []) // eslint-disable-line react-hooks/exhaustive-deps
+}
