+++ src/state/connection/constants.ts
@@ -1,0 +1,7 @@
+import { ConnectionType } from 'connectors'
+
+export const BACKFILLABLE_WALLETS = [
+  ConnectionType.COINBASE_WALLET,
+  ConnectionType.WALLET_CONNECT,
+  ConnectionType.INJECTED,
+]
